Add all files to staging area

+++ src/main/java/org/kosal/phoneshop/kosal1_phoneshop/repository/BrandRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>,JpaSpecificationExecutor<Brand>{
+public interface BrandRepository extends JpaRepository<Brand, Long>,JpaSpecificationExecutor<Brand>{
 	List<Brand>findByNameLike(String name);
 	List<Brand>findByNameContaining(String name);
